+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for beam-automigrate
 
+## 0.1.2.0
+
+* Escape sql identifiers that are on the [postgres reserved keywords list](https://www.postgresql.org/docs/current/sql-keywords-appendix.html)
+
+## 0.1.1.0
+
+* Escape sql identifiers only when required by the [postgres syntax rules](https://www.postgresql.org/docs/current/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS)
+
 ## 0.1.0.1
 
 * Loosen time version bounds